JJNotificationService/requirements.txt
# ============================
# ⚙️ Core Framework
# ============================
fastapi==0.95.2
uvicorn==0.22.0

# ============================
# 🗄️ Database
# ============================
pg8000==1.27.0
sqlalchemy==1.4.49
alembic==1.12.0

# ============================
# ✅ Validation
# ============================
pydantic==1.10.12

# ============================
# 🌐 HTTP Requests
# ============================
requests==2.31.0

# ============================
# ⚙️ Config & Environment
# ============================
python-dotenv==1.0.0

# ============================
# 📡 MikroTik API Clients
# ============================
librouteros==3.4.1           # Old method (for legacy)
routeros-api==0.18.1         # ✅ Recommended, works with RouterOS v6/v7

# ============================
# ⏰ Scheduler
# ============================
apscheduler==3.10.4

# ============================
# ⏰ For Date
# ============================
python-dateutil==2.9.0.post0
pytz==2025.1

# ============================
# 🔌 WebSockets
# ============================
websockets==11.0.3
